requirements.txt
# Flask core
Flask>=3.0.0
Werkzeug>=3.0.1

# Database
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
psycopg>=3.1.0
GeoAlchemy2>=0.14.3
Shapely>=2.0.0

# Authentication
Flask-JWT-Extended>=4.6.0
passlib[bcrypt]>=1.7.4

# API & Validation
flasgger>=0.9.7.1
marshmallow>=3.20.1
marshmallow-sqlalchemy>=0.29.0

# HTTP client (for GraphHopper)
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
httpx>=0.25.2

# Production
gunicorn>=21.2.0
